AOE Done (I hope) and Models Added
Areas of Effect for Bullets and Towers seems to be done, although the
AOE code does still need to be applied to the necessary towers.

Some models have been added to the scene, too. It is self-evident that
we need to make the models face in the correct direction as the game
runs.