Idea for Bullet Spawning at Tip
I think I have found the root of the problem with bullets not spawning
at the tip. The Instantiate function has a parameter (Vector3 position)
pertaining to the location of the copy of the object. By default, it is
set to Vector3.zero, which seems to be relative to the gameObject - the
turret. However, I can't seem to set this vector to any other value.
This problem could be solved by having a separate spawnpoint gameObject
at the tip move around with the tip, but I feel like this problem could
be easily solved by just adding a vector to move the bullet up so when
it fires it travels down the barrel. I'll look more into this
Instantiate function.