Bullets Spawn at Tip!!!
The problem was that the bullets would use an algorithm that would cause
them to travel from the tower to the target. As such, even if the bullet
were created elsewhere, the bullet would reappear in the wrong spot to
travel from the tower base to the target rather than from the tower
barrel to the target.

The Tower variable in the Bullet script now refers to a designated spawn
point that changes its own location to the end of the barrel whenever
the tower changes its own direction. It is a little fidgety when it
comes to firing at bullets closer to the center of the sniper tower than
the end of the barrel, but works well for long distances.

There are two new variables in Tower:

public GameObject bulletSpawn;
protected float bulletSpawndDistance;

bulletSpawn is the spawn point of the bullets. bulletSpawnDistance
refers to how far from the center of the tower it should be. This is
protected because it may be necessary for child Tower scripts to have
different distances since they have different models.

So it's finally done. Now all I need to do is texture the Fan Tower once
its model is put in the TowerModels folder of this repository.

- Alexander Mendelsohn