Fan Model and Texture Completed
We now have a textured model for the Fan in the TowerModels folder. Look
carefully at the way the blades are tilted, because this and whether it
should blow in or out will determine which way the fan should turn. In
the animation in Maya, I have the fan blowing out.